refactor func for sagemaker e2e
