+++ scripts/e2e_tests/aws_sagemaker/e2e_sagemaker_deployment.py
@@ -11,7 +11,10 @@ from sklearn import svm, datasets
 from bentoml import BentoService, load, api, env, artifacts
 from bentoml.artifact import PickleArtifact
 from bentoml.handlers import DataframeHandler
-
+from scripts.e2e_tests.aws_sagemaker.utils import (
+    run_sagemaker_create_or_update_command,
+    test_deployment_result,
+)
 
 logger = logging.getLogger('bentoml.test')
 
@@ -66,63 +69,12 @@ if __name__ == '__main__':
         '--num-of-gunicorn-workers-per-instance',
         '2',
     ]
-    logger.info(
-        f"Running bentoml deploy command: {' '.join(create_deployment_command)}"
+    deployment_failed, endpoint_name = run_sagemaker_create_or_update_command(
+        create_deployment_command
     )
-    with subprocess.Popen(
-        create_deployment_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE
-    ) as proc:
-        create_deployment_stdout = proc.stdout.read().decode('utf-8')
-    logger.info('Finish deploying to AWS Sagemaker')
-    logger.info(create_deployment_stdout)
-    if create_deployment_stdout.startswith('Failed to create deployment'):
-        deployment_failed = True
-    create_deployment_output_list = create_deployment_stdout.split('\n')
-    deployment_endpoint = ''
-    for index, message in enumerate(create_deployment_output_list):
-        if '"EndpointName":' in message:
-            endpoint_name = message.split(':')[1].strip(',').replace('"', '')
-            deployment_endpoint = (
-                f'https://runtime.sagemaker.{region}.amazonaws.com/'
-                f'endpoints/{endpoint_name}/invocations'
-            )
 
     if not deployment_failed and endpoint_name:
-        logger.info(f'Test deployment with sample request for {endpoint_name}')
-        try:
-            test_command = [
-                'aws',
-                'sagemaker-runtime',
-                'invoke-endpoint',
-                '--endpoint-name',
-                endpoint_name,
-                '--content-type',
-                '"application/json"',
-                '--body',
-                f'"{json.dumps(sample_data.tolist())}"',
-                '>(cat) 1>/dev/null',
-                '|',
-                'jq .',
-            ]
-            logger.info('Testing command: %s', ' '.join(test_command))
-            result = subprocess.run(
-                ' '.join(test_command),
-                capture_output=True,
-                shell=True,
-                executable='/bin/bash',
-            )
-            logger.info(result)
-            if result.stderr.decode('utf-8'):
-                logger.error(result.stderr.decode('utf-8'))
-                deployment_failed = True
-            logger.info('Prediction Result: %s', result.stdout.decode('utf-8'))
-            if '[\n  0\n]\n' == result.stdout.decode('utf-8'):
-                deployment_failed = False
-            else:
-                deployment_failed = True
-        except Exception as e:
-            logger.error(str(e))
-            deployment_failed = True
+        deployment_failed = test_deployment_result(endpoint_name, '[\n  0\n]\n')
 
     logger.info('Delete test deployment with BentoML CLI')
     delete_deployment_command = [
